Move all builtin logic to operations.cpp file
