+++ objects.cpp
@@ -23,91 +23,6 @@ TokenMap& TokenMap::default_global() {
 
 TokenMap TokenMap::empty = TokenMap(&default_global());
 
-/* * * * * TokenMap built-in functions * * * * */
-
-const args_t map_pop_args = {"key", "default"};
-packToken map_pop(TokenMap scope) {
-  TokenMap map = scope["this"].asMap();
-  std::string key = scope["key"].asString();
-
-  // Check if the item is available and remove it:
-  if (map.map().count(key)) {
-    packToken value = map[key];
-    map.erase(key);
-    return value;
-  }
-
-  // If not available return the default value or None
-  packToken* def = scope.find("default");
-  if (def) {
-    return *def;
-  } else {
-    return packToken::None;
-  }
-}
-
-packToken map_len(TokenMap scope) {
-  TokenMap map = scope.find("this")->asMap();
-  return map.map().size();
-}
-
-/* * * * * TokenList built-in functions * * * * */
-
-const args_t push_args = {"item"};
-packToken list_push(TokenMap scope) {
-  packToken* list = scope.find("this");
-  packToken* token = scope.find("item");
-
-  // If "this" is not a list it will throw here:
-  list->asList().list().push_back(*token);
-
-  return *list;
-}
-
-const args_t list_pop_args = {"pos"};
-packToken list_pop(TokenMap scope) {
-  TokenList list = scope.find("this")->asList();
-  packToken* token = scope.find("pos");
-
-  int64_t pos;
-
-  if ((*token)->type & NUM) {
-    pos = token->asInt();
-
-    // So that pop(-1) is the same as pop(last_idx):
-    if (pos < 0) pos = list.list().size()-pos;
-  } else {
-    pos = list.list().size()-1;
-  }
-
-  packToken result = list.list()[pos];
-
-  // Erase the item from the list:
-  // Note that this operation is optimal if pos == list.size()-1
-  list.list().erase(list.list().begin() + pos);
-
-  return result;
-}
-
-packToken list_len(TokenMap scope) {
-  TokenList list = scope.find("this")->asList();
-  return list.list().size();
-}
-
-/* * * * * Initialize TokenList functions * * * * */
-
-struct TokenList::Startup {
-  Startup() {
-    TokenMap& base_list = calculator::type_attribute_map()[LIST];
-    base_list["push"] = CppFunction(list_push, push_args, "push");
-    base_list["pop"] = CppFunction(list_pop, list_pop_args, "pop");
-    base_list["len"] = CppFunction(list_len, "len");
-
-    TokenMap& base_map = TokenMap::base_map();
-    base_map["pop"] = CppFunction(map_pop, map_pop_args, "pop");
-    base_map["len"] = CppFunction(map_len, "len");
-  }
-} list_startup;
 
 /* * * * * Iterator functions * * * * */
